requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
openai==1.61.*
anthropic==0.45.*
python-dotenv==1.0.*
pydantic==2.10.*
